fix(api): blockRankingsのrank/pointsをzod側スキーマ側でもintに変更

+++ apps/api/src/schemas/sportsData.ts
@@ -79,8 +79,8 @@ export const matchSchema = z.object({
 export const blockRankingSchema = z.object({
     eventBlockId: positiveIntegerSchema,
     teamId: positiveIntegerSchema,
-    rank: z.number(),
-    points: z.number(),
+    rank: z.number().int(),
+    points: z.number().int(),
     note: z.string().nullable()
 })
